+++ src/api/readme.md
@@ -1,1 +1,3 @@
-# Why do we have this versioning thing?
+# API versioning
+- This step is not necessary but it is a practice that I like to have versioning folders for APIs.
+- Following this approach we can easily support two versions of the API in the same time
