+++ src/api/v1/controllers/readme.md
@@ -1,1 +1,4 @@
-# What do we expect here?
+# Controllers Layer
+- A controller is the second layer that handles a user request.
+- Controllers layer is responsible of handling requests parameters and arguments, dispatching services and perparing responses
+- No logic should be found in this layer.
