+++ src/api/v1/middlewares/readme.md
@@ -1,1 +1,3 @@
-# What do we expect here?
+# Middleware
+- This folder is responsible of any middleware that can be used within this API such as errorHandlers, authentication, authorization and requests validation if needed.
+- For this task we will use only error handler middleware
