+++ src/api/v1/readme.md
@@ -1,1 +1,4 @@
-# What do we expect here?
+# What is expected to be found here?
+1. The service API layers, each layer would have a separated folder.
+2. Config folder contains configuration for functions and classes that is needed for the API, configured once and can be used more than one time. It is not related directly to the service domain.
+3. Each folder should have its own testing folder to test all the functions that can be found in the folder's files.
