+++ src/api/v1/routes/readme.md
@@ -1,1 +1,4 @@
-# What do we expect here?
+# Routes Layer
+- This is the first layer that the request faces in our code.
+- This layer would be responsible for assigning the suitable middleware to each single route.
+- This layer also is responsible for assigning the required controller to handle the user's request
