+++ src/api/v1/services/readme.md
@@ -1,1 +1,5 @@
-# What do we expect here?
+# Services Layer
+- This is the third layer that handles users requests.
+- This layer is the responsible of the business logic of the service.
+- Within this layer, all the service errors should be thrown.
+- Errors should be formatted using CustomError class to be more insightful.
