+++ src/api/v1/storage/readme.md
@@ -1,0 +1,4 @@
+# Services Layer
+- This is the 4th layer that handles users requests.
+- This layer is the responsible of getting data into and from th redis storage.
+- No business logic is allowed here.
