+++ src/app.ts
@@ -1,15 +1,27 @@
 import express, { Express } from "express";
 import * as dotenv from 'dotenv';
-import { errorHanlder, notFoundHandler } from "./api/v1/middlewares/ErrorHandlers";
+import { notFoundHandler } from "./api/v1/middlewares/ErrorHandlers";
 import routes from "./api/v1/routes";
+import * as Redis from "./api/v1/config/redis";
 dotenv.config();
+
+/**
+ * Closes the redis connection in case of closing the server
+ * @param app 
+ */
+export const handleCloseConnections = (app: Express): void => {
+    app.on('close', () => Redis.closeConnection());
+    process.on( 'SIGTERM', function () {
+        Redis.closeConnection()
+    });
+};
 
 /**
  * This funciton will start the express server
  * and add all the requried middleware
  * @returns 
  */
-export default () => {
+export default (): Promise<void> => {
     return new Promise((resolve) => {
 
         const app = express() as Express;
@@ -22,9 +34,10 @@ export default () => {
         app.use(notFoundHandler);
 
         const { PORT } = process.env;
+        handleCloseConnections(app);
         app.listen(PORT, () => {
             console.log(`Server is running on port ${PORT}`);
-            resolve(0);
+            resolve();
         });
     });
 }
